chore: set code coverage threshold & reorganize tests
