+++ packages/cli/src/__mocks__/fs.ts
