+++ packages/core/src/__tests__/fixtures/extension.js
