+++ packages/core/src/__tests__/fixtures/refs/definitions.yaml
