+++ packages/core/src/__tests__/fixtures/refs/examples.yaml
