+++ packages/core/src/__tests__/fixtures/refs/external-request-body.yaml
