+++ packages/core/src/__tests__/fixtures/refs/externalref.yaml
