+++ packages/core/src/__tests__/fixtures/refs/hosted.yaml
