+++ packages/core/src/__tests__/fixtures/refs/openapi-with-external-refs-conflicting-names.yaml
