+++ packages/core/src/__tests__/fixtures/refs/openapi-with-url-refs.yaml
