+++ packages/core/src/__tests__/fixtures/refs/param-b.yaml
