+++ packages/core/src/__tests__/fixtures/refs/param-c.yaml
