+++ packages/core/src/__tests__/fixtures/refs/rename.yaml
