+++ packages/core/src/__tests__/fixtures/refs/requestBody.yaml
