+++ packages/core/src/__tests__/fixtures/refs/schema-a.yaml
