+++ packages/core/src/__tests__/fixtures/refs/simple.yaml
