+++ packages/core/src/__tests__/fixtures/refs/vendor.schema.yaml
