+++ packages/core/src/__tests__/fixtures/resolve/External.yaml
