+++ packages/core/src/__tests__/fixtures/resolve/description.md
