+++ packages/core/src/__tests__/fixtures/resolve/externalInfo.yaml
