+++ packages/core/src/__tests__/fixtures/resolve/externalLicense.yaml
