+++ packages/core/src/__tests__/fixtures/resolve/openapi-with-back.yaml
