+++ packages/core/src/__tests__/fixtures/resolve/openapi-with-md-description.yaml
