+++ packages/core/src/__tests__/fixtures/resolve/openapi.yaml
