+++ packages/core/src/__tests__/fixtures/resolve/schemas/type-a.yaml
