+++ packages/core/src/__tests__/fixtures/resolve/schemas/type-b.yaml
