+++ packages/core/src/__tests__/fixtures/resolve/transitive/a.yaml
