+++ packages/core/src/__tests__/fixtures/resolve/transitive/components.yaml
