+++ packages/core/src/__tests__/normalizeVisitors.test.ts
@@ -1,7 +1,7 @@
-import { normalizeVisitors, VisitorLevelContext } from '../src/visitors';
-import { Oas3RuleSet } from '../src/oas-types';
-import { Oas3Types } from '../src/types/oas3';
-import { normalizeTypes } from '../src/types';
+import { normalizeVisitors, VisitorLevelContext } from '../visitors';
+import { Oas3RuleSet } from '../oas-types';
+import { Oas3Types } from '../types/oas3';
+import { normalizeTypes } from '../types';
 
 describe('Normalize visitors', () => {
   it('should work correctly for single rule', () => {
